Moved popup logic into an abstract class, where different popup instances can be defined
