+++ src/scripts/map/canvas.ts
@@ -1,10 +1,10 @@
 import { laneLength, metresToPixels } from "../conversions.js";
-import { getElement } from "../dom/elements.js";
-import { WAY_INFO, displayPopup } from "../dom/popup.js";
 import { DrawnElement, drawArrow, drawLine, drawPolygon, getSurfaceColour } from "../drawing.js";
+import { WAY_INFO, displayPopup } from "../popup/index.js";
 import { Settings } from "../settings.js";
 import { State } from "../state.js";
-import { zoomIncrement } from "../supplement.js";
+import { getElement } from "../supplement/elements.js";
+import { zoomIncrement } from "../supplement/index.js";
 import { ScreenCoordinate, WorldCoordinate } from "../types/coordinate.js";
 import "./buttons.js";
